Refactor plugin parser wiring into adapter
